+++ src/dataflow/getLibraryData.ts
@@ -1,11 +1,4 @@
-import {
-  OPDS2,
-  LibraryData,
-  LibraryLinks,
-  OPDS1,
-  SearchData
-} from "interfaces";
-import { OPDSFeed, OPDSShelfLink } from "opds-feed-parser";
+import { OPDS2, LibraryData, LibraryLinks, OPDS1 } from "interfaces";
 import ApplicationError, { PageNotFoundError, ServerError } from "errors";
 import { flattenSamlMethod } from "utils/auth";
 import { APP_CONFIG } from "utils/env";
@@ -24,14 +17,14 @@ async function fetchCatalogLinkBuilder(
       link => link.rel === OPDS2.CatalogLinkTemplateRelation
     )?.href;
     if (!templateUrl) {
-      throw new ApplicationError(
-        `Template not present in response from: ${registryBase}`
-      );
+      throw new ApplicationError({
+        detail: `Template not present in response from: ${registryBase}`
+      });
     }
     return uuid => templateUrl.replace("{uuid}", uuid);
   } catch (e) {
     throw new ApplicationError(
-      `Could not fetch the library template at: ${registryBase}`,
+      { detail: `Could not fetch the library template at: ${registryBase}` },
       e
     );
   }
@@ -52,50 +45,53 @@ async function fetchCatalogEntry(
     const catalogFeed = (await response.json()) as OPDS2.LibraryRegistryFeed;
     const catalogEntry = catalogFeed?.catalogs?.[0];
     if (!catalogEntry)
-      throw new ApplicationError(
-        `LibraryRegistryFeed returned by ${catalogFeedUrl} does not contain a CatalogEntry.`
-      );
+      throw new ApplicationError({
+        detail: `LibraryRegistryFeed returned by ${catalogFeedUrl} does not contain a CatalogEntry.`
+      });
     return catalogEntry;
   } catch (e) {
+    if (e instanceof ApplicationError) throw e;
     throw new ApplicationError(
-      `Could not fetch catalog entry for library: ${librarySlug} at ${registryBase}`,
+      {
+        detail: `Could not fetch catalog entry for library: ${librarySlug} at ${registryBase}`
+      },
       e
     );
   }
 }
 
-function findCatalogRootUrl(catalog: OPDS2.CatalogEntry) {
-  return catalog.links.find(link => link.rel === OPDS2.CatalogRootRelation)
+function findAuthDocUrl(catalog: OPDS2.CatalogEntry) {
+  return catalog.links.find(link => link.rel === OPDS2.AuthDocumentRelation)
     ?.href;
 }
 
 /**
- * Interprets the app config to return the catalog root url.
+ * Interprets the app config to return the auth document url.
  */
-export async function getCatalogRootUrl(librarySlug: string): Promise<string> {
+export async function getAuthDocUrl(librarySlug: string): Promise<string> {
   const libraries = APP_CONFIG.libraries;
 
   // we have a library registry url
   if (typeof libraries === "string") {
     const catalogEntry = await fetchCatalogEntry(librarySlug, libraries);
-    const catalogRootUrl = findCatalogRootUrl(catalogEntry);
-    if (!catalogRootUrl)
-      throw new ApplicationError(
-        `CatalogEntry did not contain a Catalog Root Url. Library UUID: ${librarySlug}`
-      );
-    return catalogRootUrl;
+    const authDocUrl = findAuthDocUrl(catalogEntry);
+    if (!authDocUrl)
+      throw new ApplicationError({
+        detail: `CatalogEntry did not contain a Authentication Document Url. Library UUID: ${librarySlug}`
+      });
+    return authDocUrl;
   }
   // we have a dictionary of libraries
   // just get the url from the dictionary
 
-  const catalogRootUrl =
+  const authDocUrl =
     librarySlug in libraries ? libraries[librarySlug] : undefined;
-  if (typeof catalogRootUrl !== "string") {
+  if (typeof authDocUrl !== "string") {
     throw new PageNotFoundError(
-      `No catalog root url is configured for the library: ${librarySlug}.`
+      `No authentication document url is configured for the library: ${librarySlug}.`
     );
   }
-  return catalogRootUrl;
+  return authDocUrl;
 }
 
 /**
@@ -115,33 +111,47 @@ export async function fetchAuthDocument(
 }
 
 /**
- * Extracts the loans url from a catalog root
+ * Extracts the loans url from an auth document
  */
-function getShelfUrl(catalog: OPDSFeed): string | null {
+function getShelfUrl(authDoc: OPDS1.AuthDocument): string | null {
   return (
-    catalog.links.find(link => {
-      return link instanceof OPDSShelfLink;
+    authDoc.links?.find(link => {
+      return link.rel === OPDS1.ShelfLinkRel;
     })?.href ?? null
   );
 }
 
+/**
+ * Extracts the catalot root url from an auth document
+ */
+function getCatalogUrl(authDoc: OPDS1.AuthDocument): string {
+  const url =
+    authDoc.links?.find(link => {
+      return link.rel === OPDS1.CatalogRootRel;
+    })?.href ?? null;
+
+  if (!url)
+    throw new ApplicationError({
+      detail: "No Catalog Root Url present in Auth Document."
+    });
+
+  return url;
+}
 /**
  * Constructs the internal LibraryData state from an auth document,
  * catalog url, and library slug.
  */
 export function buildLibraryData(
   authDoc: OPDS1.AuthDocument,
-  catalogUrl: string,
-  librarySlug: string,
-  catalog: OPDSFeed,
-  searchData?: SearchData
+  librarySlug: string
 ): LibraryData {
   const logoUrl = authDoc.links?.find(link => link.rel === "logo")?.href;
   const headerLinks =
     authDoc.links?.filter(link => link.rel === "navigation") ?? [];
   const libraryLinks = parseLinks(authDoc.links);
   const authMethods = flattenSamlMethod(authDoc);
-  const shelfUrl = getShelfUrl(catalog);
+  const shelfUrl = getShelfUrl(authDoc);
+  const catalogUrl = getCatalogUrl(authDoc);
   return {
     slug: librarySlug,
     catalogUrl,
@@ -157,8 +167,7 @@ export function buildLibraryData(
         : null,
     headerLinks,
     libraryLinks,
-    authMethods,
-    searchData: searchData ?? null
+    authMethods
   };
 }