+++ src/errors.ts
@@ -2,53 +2,57 @@ import { OPDS1 } from "interfaces";
 
 export default class ApplicationError extends Error {
   info: OPDS1.ProblemDocument;
+  baseError?: Error;
 
-  constructor(m: string, baseError?: Error) {
-    super(`${m}${baseError ? `\nBase Error:\n${baseError.message}` : ""}`);
-    Object.setPrototypeOf(this, ApplicationError.prototype);
-    this.name = "Application Error";
-    this.info = {
+  constructor(info: Partial<OPDS1.ProblemDocument>, baseError?: Error) {
+    const problemDoc = {
       title: "Application Error",
-      detail: m
+      detail: "An unknown Application Error Occurred",
+      ...info
     };
+    super(`${problemDoc.title}: ${problemDoc.detail}`);
+    Object.setPrototypeOf(this, ApplicationError.prototype);
+    this.name = `ApplicationError${info.title ? `: ${info.title}` : ""}`;
+    this.baseError = baseError;
+    this.info = problemDoc;
   }
 }
 
 export class PageNotFoundError extends ApplicationError {
   constructor(m: string, baseError?: Error) {
-    super(`${m}${baseError ? baseError.message : ""}`);
-    Object.setPrototypeOf(this, PageNotFoundError.prototype);
-    this.name = "Page Not Found Error";
-    this.info = {
+    const info = {
       title: "Page Not Found",
       detail: m,
       status: 404
     };
+    super(info, baseError);
+    Object.setPrototypeOf(this, PageNotFoundError.prototype);
+    this.name = "Page Not Found Error";
   }
 }
 
 export class UnimplementedError extends ApplicationError {
   constructor(m: string, baseError?: Error) {
-    super(`${m}${baseError ? baseError.message : ""}`);
-    Object.setPrototypeOf(this, UnimplementedError.prototype);
-    this.name = "Unimplemented Error";
-    this.info = {
+    const info = {
       title: "Unimplemented",
       detail: m
     };
+    super(info, baseError);
+    Object.setPrototypeOf(this, UnimplementedError.prototype);
+    this.name = "Unimplemented Error";
   }
 }
 
 export class AppSetupError extends ApplicationError {
   constructor(m: string, baseError?: Error) {
-    super(`${m}${baseError ? baseError.message : ""}`);
-    Object.setPrototypeOf(this, AppSetupError.prototype);
-    this.name = "App Setup Error";
-    this.info = {
+    const info = {
       status: 500,
       title: "App Setup Error",
       detail: m
     };
+    super(info, baseError);
+    Object.setPrototypeOf(this, AppSetupError.prototype);
+    this.name = "App Setup Error";
   }
 }
 
@@ -59,15 +63,14 @@ export class FetchError extends ApplicationError {
   url: string;
 
   constructor(url: string, baseError?: Error) {
-    super(`${baseError ? baseError.message : ""}`);
+    const info = {
+      title: "Fetch Error",
+      detail: `The fetch promise for the requested resource was rejected. This is probably an offline, CORS, or other network error. Requested URL: ${url}`
+    };
+    super(info, baseError);
     Object.setPrototypeOf(this, FetchError.prototype);
     this.name = "Fetch Error";
     this.url = url;
-    this.info = {
-      title: "Fetch Error",
-      detail:
-        "The fetch promise for the requested resource was rejected. This is probably an offline, CORS, or other network error."
-    };
   }
 }
 
@@ -79,11 +82,6 @@ function isProblemDocument(
 export class ServerError extends ApplicationError {
   // a default problem document
   url: string;
-  info: OPDS1.ProblemDocument = {
-    detail: "An unknown error server occurred.",
-    title: "Server Error",
-    status: 500
-  };
   authDocument?: OPDS1.AuthDocument;
 
   constructor(
@@ -91,21 +89,33 @@ export class ServerError extends ApplicationError {
     status: number,
     details: OPDS1.ProblemDocument | OPDS1.AuthDocument
   ) {
-    super("Server Error");
-    this.url = url;
-    Object.setPrototypeOf(this, ServerError.prototype);
-    if (status === 401 && !isProblemDocument(details)) {
-      // 401 errors return auth document instead of problem document
-      // we will construct our own problem document.
-      this.info = {
+    // default info in case no problem doc came back and it wasn't a 401
+    let info: OPDS1.ProblemDocument = {
+      title: "Unknown Server Error",
+      detail: `The Circulation Manager returned a ${status} error for: ${url}.`,
+      status
+    };
+    if (isProblemDocument(details)) {
+      // if the server returned a problem document
+      info = details;
+    } else if (status === 401) {
+      // if the server returned a 401 with an authentication document
+      // we build our own problem document
+      info = {
         title: "Not Authorized",
         detail: "You are not authorized for the requested resource.",
         status: 401
       };
-      this.authDocument = details;
-    } else if (isProblemDocument(details)) {
-      this.info = details;
     }
+    super(info);
+
+    // if the server returned a 401 with a non problem document, we assume
+    // it returned an authentication document.
+    if (status === 401 && !isProblemDocument(details))
+      this.authDocument = details;
+
+    this.url = url;
+    Object.setPrototypeOf(this, ServerError.prototype);
     this.name = `Server Error`;
   }
 }