+++ src/interfaces.ts
@@ -96,7 +96,6 @@ export interface LibraryData {
   libraryLinks: LibraryLinks;
   authMethods: AppAuthMethod[];
   shelfUrl: string | null;
-  searchData: SearchData | null;
 }
 
 /**
@@ -229,6 +228,7 @@ export interface CollectionData {
   parentLink?: LinkData | null;
   shelfUrl?: string;
   links?: LinkData[] | null;
+  searchDataUrl: string | null;
   raw?: any;
 }